Refactor client id from byte to ushort

+++ scripts/mob/player/Player.cs
@@ -24,7 +24,7 @@ public partial class Player : Mob
 
 #nullable enable
     public UdpClient? m_UdpClient;
-    public byte? m_ClientId { get; private set; }
+    public ushort? m_ClientId { get; private set; }
 
     // Aiming/Camera input
     private float m_YawInput = 0.0f;
@@ -152,7 +152,7 @@ public partial class Player : Mob
         m_CameraPivot.Rotation = cameraRotation;
     }
 
-    public void SetClientId(byte? clientId)
+    public void SetClientId(ushort? clientId)
     {
         if (clientId == null) { m_ClientId = null; }
         else { m_ClientId = clientId; }
